refactor(pre-planner): decouple planner import
